bản sửa 9/3/2018

+++ app/src/main/java/com/vnpt/vnptline/domain/model/pojo/response/hotel/RoomAvailableResponse.java
@@ -79,7 +79,7 @@ public class RoomAvailableResponse {
     @SerializedName("priceHour")
     @Expose
     @Setter @Getter
-    public double priceHour;
+    public Integer priceHour;
 
     @SerializedName("promotionPriceHour")
     @Expose
